finishes adding voting feature

+++ app/src/main/java/com/dangerfield/wouldyourather/model/Question.kt
@@ -1,6 +1,6 @@
 package com.dangerfield.wouldyourather.model
 
-data class Question(var option1Votes: Long, var option2Votes: Long, val questions: ArrayList<String>) {
+data class Question(var option1Votes: Double, var option2Votes: Double, val questions: ArrayList<String>) {
 
-    constructor(): this(0.toLong(),0.toLong(),ArrayList<String>())
+    constructor(): this(0.toDouble(),0.toDouble(),ArrayList<String>())
 }
